+++ src/features/bible/index.ts
@@ -1,3 +1,4 @@
 export * from './api';
 export * from './hooks';
 export * from './lib';
+export * from './model';
