+++ src/features/bible/model/constants.ts
