+++ src/features/bible/model/index.ts
