+++ src/features/bible/model/types.ts
