+++ src/features/news/api/fetchRssFeed.ts
@@ -1,7 +1,7 @@
 import type { RSSFeed, RSSItem } from '../model';
 import axios from 'axios';
 import Parser from 'rss-parser';
-import { extractThumbnail } from '..';
+import { extractThumbnail } from '@/features/news/lib';
 import { extractLastNumber } from '@/shared';
 
 export async function fetchRssFeed(feedUrl: string, sourceName: Record<string, string>) {
