+++ src/features/news/index.ts
@@ -1,1 +1,3 @@
 export * from './api';
+export * from './lib';
+export * from './model';
