+++ src/features/news/lib/index.ts
@@ -1,2 +1,1 @@
 export * from './extractThumbnail';
-export * from './fetchRssFeed';
