+++ src/features/news/model/index.ts
