+++ src/features/news/model/types.ts
